+++ src/main/java/com/pi4j/drivers/display/graphics/awt/JFrameGraphicsDriver.java
@@ -1,4 +1,4 @@
-package com.pi4j.drivers.display.graphics.jframe;
+package com.pi4j.drivers.display.graphics.awt;
 
 import com.pi4j.drivers.display.graphics.GraphicsDisplayDescriptor;
 import com.pi4j.drivers.display.graphics.GraphicsDisplayDriver;
